+++ extensions/forked-extensions/src/components/diagnostics.tsx
@@ -3,6 +3,7 @@ import { Action, ActionPanel, Detail, useNavigation } from "@raycast/api";
 import * as api from "../api.js";
 import { catchError } from "../errors.js";
 import * as git from "../git.js";
+import { getCommitDiffMessage, repositoryConfigurationPath } from "../utils.js";
 
 export default function Diagnostics() {
   const [status, setStatus] = useState<string>("Running diagnostics...");
@@ -19,25 +20,56 @@ export default function Diagnostics() {
       const remoteForkedRepository = localForkedRepository
         ? await api.repositoryExists(localForkedRepository)
         : "✅ Cannot determine (no local repository found)";
+
+      const githubCommitDiff = localForkedRepository ? await api.compareTwoCommits(localForkedRepository) : undefined;
+      const githubCommitDiffPass = githubCommitDiff && !(githubCommitDiff.ahead > 0 || githubCommitDiff.behind > 0);
+      const githubCommitDiffMessage = getCommitDiffMessage(githubCommitDiff, {
+        includeAhead: true,
+        includeZeroAhead: true,
+        alwaysShow: true,
+      }).trim();
+      const githubCommitDiffGuide = githubCommitDiffPass ? "" : ' (You can use "Sync Remote" action to sync changes)';
+
+      const localCommitDiff = await git.getAheadBehindCommits();
+      const localCommitDiffPass = !(localCommitDiff.ahead > 0 || localCommitDiff.behind > 0);
+      const localCommitDiffMessage = getCommitDiffMessage(localCommitDiff, {
+        includeAhead: true,
+        includeZeroAhead: true,
+        alwaysShow: true,
+      }).trim();
+      const localCommitDiffGuide = localCommitDiffPass ? "" : ' (You can use "Pull Changes" action to sync changes)';
+
       const status = [
         "## Diagnostics",
         "### Git installed",
         isGitInstalled
-          ? "✅ Good"
-          : "⚠️ Git is not installed or not found. Please set up your Git executable file path manually in the extension preferences.",
+          ? "- ✅ Good"
+          : "- ⚠️ Git is not installed or not found. Please set up your Git executable file path manually in the extension preferences.",
         "### Git status clean",
         isStatusClean
-          ? "✅ Good"
-          : "⚠️ You have uncommitted changes. Please commit or stash them before performing operations.",
+          ? "- ✅ Good"
+          : "- ⚠️ You have uncommitted changes. Please commit or stash them before performing operations.",
         "### Local forked repository",
         localForkedRepository
-          ? `✅ ${localForkedRepository}`
-          : "⚠️ Not found, please rerun the extension to re-initialize the repository.",
+          ? [
+              `- ✅ [${localForkedRepository} 📁](file://$${repositoryConfigurationPath})`,
+              `- ${localCommitDiffPass ? "✅" : "⚠️"} ${localCommitDiffMessage}${localCommitDiffGuide}`,
+            ]
+              .filter(Boolean)
+              .join("\n")
+          : "- ⚠️ Not found, please rerun the extension to re-initialize the repository.",
         "###  Remote forked repository",
         remoteForkedRepository === true
-          ? `✅ ${localForkedRepository}`
+          ? [
+              `- ✅ [${localForkedRepository} 🌐](https://github.com/${localForkedRepository})`,
+              localForkedRepository
+                ? `- ${githubCommitDiffPass ? "✅" : "⚠️"} ${githubCommitDiffMessage}${githubCommitDiffGuide}`
+                : "",
+            ]
+              .filter(Boolean)
+              .join("\n")
           : remoteForkedRepository === false
-            ? "⚠️ Not found, please check if the repository still exists on GitHub or if your GitHub token has the necessary permissions."
+            ? "- ⚠️ Not found, please check if the repository still exists on GitHub or if your GitHub token has the necessary permissions."
             : remoteForkedRepository,
       ].join("\n\n");
       setStatus(status);